+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.5.2
 
+* Synchronizes `VideoPlayerValue.isPlaying` with underlying video player.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.5.1