+++ packages/video_player/video_player_web/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.0.14
 
+* Synchronizes `VideoPlayerValue.isPlaying` with `VideoElement`.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.0.13